Thu, Sep 15, 2022, 7:59 PM +00:00

+++ readability/readability.c
@@ -27,7 +27,7 @@ int main(void)
 	int sentences = count_sentences(text);
 	printf("Number of sentences: %i\n", sentences);
 
-	float X = float(length/words);
+	float X = (length/words);
 	printf("X: %f\n", X);
